% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idates.R
\name{addPeriods}
\alias{addPeriods}
\title{Add a number of periods in a certain frequency to an IDate}
\usage{
addPeriods(x, n, frequency)
}
\arguments{
\item{x}{The IDate}

\item{n}{The number of periods to add}

\item{frequency}{Integer representing the frequency of the n periods to add}
}
\description{
Add a number of periods in a certain frequency to an IDate
}
