% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idates.R
\name{as.IDates}
\alias{as.IDates}
\title{Convert a list of string on the format YYYY-MM-DD to a list of IDate}
\usage{
as.IDates(x)
}
\arguments{
\item{x}{The list of strings}
}
\value{
A list of IDate
}
\description{
Convert a list of string on the format YYYY-MM-DD to a list of IDate
}
