% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{day_of_week}
\alias{day_of_week}
\title{Helper function to calculate the day of the week (1 = Monday, ..., 7 = Sunday)}
\usage{
day_of_week(year, month, day)
}
\description{
Helper function to calculate the day of the week (1 = Monday, ..., 7 = Sunday)
}
