% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idates.R
\docType{class}
\name{IDate-class}
\alias{IDate-class}
\title{IDate class, a date represented by three integers, Year, Month and Day.}
\description{
IDate class, a date represented by three integers, Year, Month and Day.
}
