\name{idates-package}
\alias{idates-package}
\alias{idates}
\docType{package}
\title{
\packageTitle{idates}
}
\description{
\packageDescription{idates}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{idates}

Maintainer: \packageMaintainer{idates}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = idates)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{idates}}
%\section{Documentation Index}{\packageIndices{idates}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
