% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idates.R
\name{startOfPeriod}
\alias{startOfPeriod}
\title{Set an IDate to the start of period}
\usage{
startOfPeriod(x, frequency)
}
\arguments{
\item{x}{The IDate}

\item{frequency}{The frequency to set the IDate to the start of period of}
}
\description{
Set an IDate to the start of period
}
