% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idates.R
\name{toString,IDate-method}
\alias{toString,IDate-method}
\title{S4 Method to convert an IDate to string for printing or further parsing}
\usage{
\S4method{toString}{IDate}(x, ...)
}
\description{
S4 Method to convert an IDate to string for printing or further parsing
}
