% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idates.R
\name{toString.IDate}
\alias{toString.IDate}
\title{S3 Method to convert an IDate to string for printing or further parsing, calls the S4 method}
\usage{
\method{toString}{IDate}(x, ...)
}
\description{
S3 Method to convert an IDate to string for printing or further parsing, calls the S4 method
}
